cleanup playground and prepare it for native support
