+++ playground/src/jvmMain/kotlin/TcpServerApp.kt
@@ -14,20 +14,6 @@
  * limitations under the License.
  */
 
-package tcp
-
-import io.ktor.network.selector.*
-import io.ktor.network.sockets.*
-import io.ktor.util.*
-import io.rsocket.kotlin.core.*
 import kotlinx.coroutines.*
-import rSocketAcceptor
-import java.util.concurrent.*
 
-@OptIn(KtorExperimentalAPI::class)
-fun main(): Unit = runBlocking {
-    val dispatcher = Executors.newCachedThreadPool().asCoroutineDispatcher()
-    val server = aSocket(ActorSelectorManager(dispatcher)).tcp().bind("127.0.0.1", 2323)
-
-    server.rSocket(acceptor = rSocketAcceptor)
-}
+suspend fun main(): Unit = runTcpServer(Dispatchers.IO)
